Modificacion de model a entity - mapeo con incorporacion de anotaciones (#96)

+++ TP2_Grupo12/src/main/java/ar/edu/unju/fi/service/IConsejoService.java
@@ -2,7 +2,7 @@ package ar.edu.unju.fi.service;
 
 import java.util.List;
 
-import ar.edu.unju.fi.model.Consejo;
+import ar.edu.unju.fi.entity.Consejo;
 import jakarta.validation.Valid;
 
 public interface IConsejoService {
@@ -11,7 +11,7 @@ public interface IConsejoService {
 	
 	void guardar(@Valid Consejo consejo);
 	
-	Consejo buscar(int num);
+	Consejo buscar(Long num);
 	
 	void modificar(Consejo consejo);
